==> R/02_reporting.qmd <==
---
title: "Results Overview"
author: "Dave"
format:
  docx:
    fig-dpi: 600
    fig-width: 6
    reference-doc: caru-template.docx
editor: visual
echo: false
# knitr: 
#   opts_chunk: 
#     dev: "cairo_svg"
---

# Hello!

Welcome to this overview of SCPN quant data, which provides graphs and tabulations of most of the questions from our survey. If you're looking for totals and averages, check the table right at the bottom.

## Set up

### Libraries

```{r libraries}
library(fs)
library(here)
library(ggplot2)
library(dplyr)
library(carutools)
library(forcats)
library(scales)
library(stringr)
library(knitr)
library(lubridate)
library(tidyr)
library(purrr)
library(ggbeeswarm)
library(rlang)
```

### Read data

```{r read_data}
scot <- readRDS(here("data", "output-01_cleaned-formatted.RDS"))

scot <- filter(scot, !denomination_other %in% "Test sample")
```

## Helpers

```{r helpers}
theme_ca <- theme_ca("cyan") + theme(axis.ticks = element_line())

theme_set(theme_ca)

update_geom_defaults("bar", aes(fill = ca_cyan(), colour = "black"))


percentify <- \(x, prec = 1) label_percent(prec)(prop.table(x))

add_percent <- function(x, col = n){
  col <- enexpr(col)
  
  mutate(x, `%` = percentify(!!col))
}

count_with_percent <- function(x, group){
  
  group = enexpr(group)
  
  group_by(x, !!group) |> 
    count() |> 
    ungroup() |> 
    arrange(-n, !!group) |> 
    add_percent()
  
}

factor_table <- function(col, x = scot){
  
  col <- enexpr(col)
  
  count_with_percent(x, !!col) |> 
    rename_with(str_to_title, 1)
}

church_plants <- \(x) rename(x, `Church plants` = n)

relabeller <- function(...){
  
  dots <- enexprs(...)
  
  new_function(
    args = pairlist2(x =),
    body = expr({
      case_match(
        x,
        !!!dots,
        .default = x
      )}
      ))}
```

## Reporting

### Denomination

What denomination/network/movement is your church plant affiliated with?

```{r denomination}
denom <- 
mutate(scot, denomination =
         ordered(denomination) |> 
         fct_infreq() |> 
         fct_lump_n(7)) |>
  
  filter(!is.na(denomination)) 
  

ggplot(denom, aes(x = denomination)) +
  
  geom_bar() +
  
  ylab("Church plants") + 
  xlab("Denomination") +
  
  scale_x_discrete(
    labels = relabeller(
      "Assemblies of God" ~ "Assemblies of\nGod",
      "Scottish Network Churches" ~ "Scottish Network\nChurches",
      "Church of Scotland" ~ "CoS",
      "New Frontiers" ~ "New\nFrontiers",
      "Free Church of Scotland" ~ "Free CoS",
      "Redeemed Christian Church of God" ~ "RCCoG"
    )
  ) +
  
  scale_y_continuous(
    sec.axis = sec_axis(
      transform = \(x) x / sum(!is.na(scot$denomination)),
      labels = label_percent(1),
      breaks = seq(0, 20, 5)/ sum(!is.na(scot$denomination)))
    ) + 
  
  ggtitle("Denominations of responding church plants") +
  
  theme(panel.grid.major.x = element_blank())

denom <-
factor_table(denomination,
             filter(scot, !is.na(denomination))) |> 
  church_plants() |> 
  set_names("Denomination", "Church plants (all time)", "% (all time)")

denom_2010 <-
  factor_table(
    denomination,
    filter(scot, !is.na(denomination), start_year >= 2010)) |> 
  church_plants() |> 
  set_names("Denomination", "Church plants (since 2010)", "% (since 2010)")

left_join(denom, denom_2010, by = "Denomination") |> 
  kable()
```

## Launch date

When did the church plant start?

```{r launch_date}
scot <- mutate(scot, launch_date = make_date(start_year, start_month))

dates <- 
  group_by(scot, launch_date) |> 
  count() |> 
  ungroup() |> 
  arrange(launch_date) |> 
  mutate(n_plants = cumsum(n))

ggplot(dates, aes(x = launch_date, y = n_plants)) + 
  geom_rect(aes(ymin = 0,
                xmax = lead(launch_date,
                           default = max(launch_date) + period(4, "month")),
                xmin = launch_date,
                ymax = n_plants),
            fill = ca_cyan(),
            alpha = 0.7) + 
  
  geom_step(linewidth = 1, colour = ca_cyan()) +
  
  
  
  scale_y_continuous(
    name = "Church plants", 
    limits = c(0, 90),
    breaks = 0:10 * 10
  ) +
  scale_x_date(name = "Date") +
  ggtitle("Cumulative church plants over time")

group_by(scot, start_year) |> 
  count(name = "Churches planted") |> 
  ungroup() |> 
  arrange(start_year) |> 
  filter(!is.na(start_year), start_year > 1900) |> 
  rename(`Launch year` = start_year) |> 
  complete(`Launch year` = full_seq(`Launch year`, 1), 
           fill = list(`Churches planted` = 0)) |> 
  kable()

cppy_mean<- 
  filter(scot, start_year >= 2010) |> 
  group_by(start_year) |> 
  count() |>
  pull(n) |> 
  mean()

ggplot(scot |> filter(start_year >= 2010),
       aes(x = start_year)) + 
    
  geom_hline(yintercept = cppy_mean, linetype = "dashed") + 
  geom_bar(alpha = 0.9) +
  
  annotate("text", label = "Yearly average", 
           x = 2011, y = cppy_mean + 0.5, size = 4) +
  
  scale_y_continuous(name = "Churches planted",
                     breaks = seq(0, 10, 2)) +

  
  xlab("Year of Launch") +
  ggtitle("Responding church plants by launch year",
          "Since 2010")
```

### Attenders (current and at launch)

Questions asked:

-   What average attendance (or sized team) did you launch with? \[Adults \| Children\]

-   What is your current average attendance \[Adults \| Children\]

#### Summary table

Some measures of central tendency to begin with

```{r attenders}

attenders <- 
  scot |> 
  select(contains("attenders"), respondent_id) |>
  pivot_longer(-respondent_id,
               names_to = "stat") |> 
  separate_wider_regex(stat,
                       patterns = c(
                         time = "launch|current",
                         "_attenders_",
                         age  = "children|adult"
                         )) |> 
  pivot_wider(names_from = time, values_from = value) |> 
  unnest(everything())

attenders_summary <- 
  group_by(attenders, age) |> 
  select(-respondent_id) |> 
  summarise(
    across(everything(),
           .fn = list(
             mean   = \(x) mean(x, na.rm = TRUE),
             median = \(x) median(x, na.rm = TRUE),
             max    = \(x) max(x, na.rm = TRUE),
             min    = \(x) min(x, na.rm = TRUE)), 
           .names = "{.fn} ({.col})"
           )) |> 
  mutate(across(where(is.numeric), \(x) round(x, 2)))

attenders_summary <- attenders_summary[order(names(attenders_summary))]

kable(attenders_summary)
```

### Lollipop

```{r lollipop}
#| fig-width: 6

attenders <- mutate(attenders,
                    change = current - launch,
                    age = str_to_title(age) |> 
                      str_replace("Adult", "Adults"))

positions <- 
  summarise(attenders, change = sum(change, na.rm = TRUE), .by = respondent_id) |> 
  arrange(change) |> 
  mutate(pos = row_number()) |> 
  select(respondent_id, pos)

attenders <- left_join(attenders, positions, by = "respondent_id")

long_tenders <- pivot_longer(attenders, c(launch, current),
                             names_to = "time", values_to = "attenders")

ggplot(long_tenders, aes(x = factor(pos))) +
  
    
  geom_segment(mapping = aes(y = launch, yend = current),
               data = attenders,
               linewidth = 1,
               alpha = 0.2) +

  geom_point(data = long_tenders, aes(y = attenders, colour = time),
             size = 1, alpha = 0.7) +
  
  scale_colour_manual(values = c(ca_cyan(), ca_purple()),
                      name = "Time point") +
  
  xlab("Churches") +
  ylab("Attenders") +
  
  theme(axis.ticks.x = element_blank(),
        panel.grid.major.x = element_blank(),
        axis.text.x = element_blank()) +
  
  facet_wrap(~ age, dir = "h") +
  ggtitle("Attendance has increased at most church plants",
          sub = "For adults and children alike")
```

### Current attenders

```{r current_attenders}
n_resp_att <- ceiling(sum(!is.na(attenders$current))/2)

attenders |> 
  mutate(age = str_to_title(age)) |> 
ggplot(aes(x = current)) +
  geom_histogram(fill = ca_cyan(),
                 binwidth = 10, boundary = 0) +
  scale_x_continuous(breaks = seq(0, 400, 20))+
  xlab("Attenders") +
  ylab("Church plants") + 
  
  scale_y_continuous(
    sec.axis = 
      sec_axis(
        transform = \(x) x / n_resp_att,
        labels = label_percent(1),
        breaks = seq(0, 40, 10)/n_resp_att
      )
  ) +
  
  facet_wrap(~ age, dir = "v") +
  ggtitle("Current attenders")


attenders |>
  summarise(attenders = sum(current), .by = respondent_id) |> 
  
  ggplot(aes(x = attenders)) +
  geom_histogram(fill = ca_cyan(), binwidth = 15, boundary = 0.9) +
  scale_x_continuous(name = "Attenders", breaks = seq(0, 200, 30),
                     limits = c(0, 200)) +
  scale_y_continuous(
    sec.axis = 
      sec_axis(
        transform = \(x) x / n_resp_att,
        labels = label_percent(1),
        breaks = seq(0, 30, 5)/n_resp_att
      )) +
  
  ylab("Church plants") +
  ggtitle("Current attenders")

## total attenders:

sum(attenders$current)
sum(attenders$launch)
```

### Plant type

Which of the following best describes how you church plant started?

-   Relaunch or revitalization of an existing church (church graft or transplant)

-   An entirely new church in a new location (physically separate from sending church)

-   Establishment of a new worshiping community linked to an existing church

```{r plant_type}

n_cpt_resp <- sum(!is.na(scot$plant_type))

scot |> 
  filter(!is.na(plant_type)) |> 
ggplot(aes(x = plant_type)) +
  geom_bar() +
  scale_x_discrete(name = "Church plant type", drop = FALSE,
                   labels = 
                     relabeller(
                       "Entirely new church" ~ "Entirely new\nchurch",
                       "Linked worshipping community" ~
                         "Linked worshipping\ncommunity"
                     )) +
  ylab("Church plants") +
  ggtitle("Types of church plant") +
  scale_y_continuous(breaks = seq(0, 100, 10),
                     sec.axis = 
                       sec_axis(transform = \(x) x/ n_cpt_resp,
                                labels    = label_percent(1),
                                breaks = seq(0, 100, 10)/n_cpt_resp)) +
  theme(panel.grid.major.x = element_blank())

filter(scot, !is.na(plant_type)) |> 
count_with_percent(plant_type) |> 
  rename(`Plant type` = plant_type) |>
  church_plants() |> 
  kable()
```

## Geography

Which of these best describes the geographical area you planted into? Please tick all that apply

```{r geography}
not_na <- \(x) !is.na(x)

geog_responders <- 
  filter(scot, if_any(starts_with("geography_"), not_na)) |> 
  nrow()


geogs <- 
  scot |> 
  pivot_longer(starts_with("geography_") & !ends_with("_other"),
               names_prefix = "geography_",
               names_to = "geography", values_to = "is") |>
  mutate(geography =
           str_to_title(geography) |> 
           str_replace_all("_", " ") |> 
           ordered()) |> 
  summarise(plants = sum(is), .by = geography) |> 
  mutate(geography =
           fct_lump_n(geography, 6, w = plants) |> 
           fct_infreq(w = plants) |> 
           fct_relevel("Other", after = Inf)) |> 
  summarise(plants = sum(plants), .by = geography)

geog_table <-
  geogs |>
  arrange(-plants) |> 
  mutate(`% responding plants` = label_percent(1)(plants / geog_responders)) 


geog_table |> 
  filter(plants > 0) |> 
  rename(`Church plants` = plants) |> 
  kable()

geog_table |> 
  filter(plants > 0) |> 
  ggplot(aes(x = geography, y = plants)) +
  geom_col() +
  scale_x_discrete(name = "Geographical area",
                   labels = 
                     relabeller(
                       "Specific geographical area within a larger town or city" ~
                         "Specific area\nin town/city"
                     )) +
  ggtitle("Geographical areas into which churches planted",
         sub = "(Non-exclusive categories)") +
  ylab("Church plants") +
  scale_y_continuous(
    sec.axis = sec_axis(
      transform = \(x) x/geog_responders,
      labels = label_percent(1),
      breaks = seq(0, 40, 10)/geog_responders)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

### Traditions

What were the traditions of the church planting team? Tick all that apply

```{r traditions}
trad_responders <- 
  filter(scot, if_any(starts_with("tradition_"), not_na)) |> 
  nrow()

trads <- 
  scot |> 
  pivot_longer(starts_with("tradition_") & !ends_with("_other"),
               names_prefix = "tradition_",
               names_to = "tradition", values_to = "is") |>
  summarise(plants = sum(is), .by = tradition)

trad_table <-
  trads|>
  arrange(-plants) |> 
  mutate(`% responding plants` = label_percent(1)(plants / trad_responders)) |> 
  mutate(tradition =
           str_to_title(tradition) |> 
           str_replace_all("_", " ") |> 
           ordered() |> 
           fct_inorder())

trad_table |>
  rename(`Church plants` = plants) |>
  kable()

trad_table |> 
  filter(plants > 0) |> 
  ggplot(aes(x = tradition, y = plants)) +
  geom_col() +
  scale_x_discrete(name = "Tradition", guide = guide_axis(n.dodge = 2)) +
  ggtitle("Traditions of church-planting teams",
         sub = "(Non-exclusive categories)") +
  ylab("Church plants") +
  scale_y_continuous(
    breaks = seq(0, 100, 10),
    sec.axis = sec_axis(transform = \(x) x/trad_responders,
                        labels = label_percent(1),
                        breaks = seq(0, 100, 10)/trad_responders)
  )
```

### Location

Where dioes the church plant meet:

-   In someone's home or a variety of homes

-   In another church's building, which we rent

-   In another church's building, which we use free of charge (this includes the sending church's building)

-   In a rented space that isn't a church building

-   In a space owned exclusively by the church plant

-   Other (please specify)

```{r location}
filter(scot, !is.na(location)) |> 
  count_with_percent(location) |> 
  church_plants() |> 
  kable()

n_loc_r <- sum(!is.na(scot$location))

ggplot(filter(scot, !is.na(location)), aes(x = location)) + 
  
  geom_bar() +
  
  ylab("Church plants") + 
  xlab("Location") +

  scale_y_continuous(
    sec.axis = sec_axis(
      transform = \(x) x / n_loc_r,
      labels = label_percent(1),
      breaks = seq(0, 50, 10) / n_loc_r
    )
  ) +
  
  scale_x_discrete(
    guide = guide_axis(n.dodge = 1),
    labels = 
      relabeller(
        "Rented space (not church building)" ~ 
          "Rented space\n(not church building)",
        "Rented church building" ~
          "Rented church\nbuilding",
        "Free church building" ~
          "Free church\nbuilding",
        "Space owned by plant" ~
          "Space owned\nby plant"
      )) +
  
    ggtitle("Meeting locations used by church plants")
```

"Other" responses:

```{r other_location_responses}
#| output: asis
scot$location_other[!is.na(scot$location_other)] |> 
  str_c("* ", foo = _,  sep = "", collapse = "\n") |> 
  cat()
```

## FTE

Please use the sliders below to indicate the percentage of full time hours worked for the church plant/new worshiping community by any paid members of staff. If a staff member splits their role between the church plant and other responsibilities, please estimate the amount of their paid time that they spend working for the church plant.

#### Members of staff

```{r staff_members}
add <- function(x){
  x[is.na(x)] <- 0
  reduce(x, `+`)
}

employees <- 
  select(scot, respondent_id, starts_with(c("percent_fte", "staff_member"))) |> 
  mutate(across(-respondent_id, not_na)) |>
  pivot_longer(-respondent_id) |> 
  summarise(employees = sum(value), .by = respondent_id)

scot <- left_join(scot, employees, by = "respondent_id")

ggplot(scot, aes(x = employees)) + 
  geom_histogram(binwidth = 1) +
  
  scale_x_continuous("Paid members of staff", breaks = 0:8) +
  scale_y_continuous(
    sec.axis = sec_axis(
      transform = \(x) x/nrow(scot),
      labels = label_percent(1),
      breaks = seq(0, 40, 10)/nrow(scot))) +
  
  ylab("Church plants") +
  ggtitle("Paid members of staff across all church plants") 

count_with_percent(scot, employees) |> 
  arrange(employees) |> 
  church_plants() |> 
  kable()
```

#### Full time equivalent hours worked

```{r fte_worked}
beeswarmify <- function(x, mult = 1){
  len <- length(x)
  pos <- 1:len
  
  (pos - mean(pos)) * mult
}

scot <-
  rowwise(scot) |> 
  mutate(fte = add(c_across(starts_with(c("percent_fte_", "staff_member_"))))) |>
  mutate(fte = fte/100) |> 
  ungroup()

# scot$fte

foo <- 
ggplot(scot) +
  geom_density(aes(x = fte))

bar <- 
ggplot(scot) + 
  geom_point(aes(x = employees, y = fte), 
             alpha = 0.5, colour = ca_cyan(), size = 2) +
  geom_abline(intercept = 0, slope = 0.5, 
              linetype = "dashed")

# scot |> 
#   pivot_longer(starts_with(c("percent_fte_", "staff_member")),
#                names_to = "remove", values_to = "employee_fte") |> 
#   select(-remove) |> 
#   mutate(respondent_id = ordered(respondent_id)) |> 
#   filter(!is.na(employee_fte)) |> 
#   mutate(employee_fte = employee_fte/100,
#          respondent_id = fct_reorder(respondent_id, -employee_fte, sum)) |>
#   mutate(y_pos = beeswarmify(employee_fte, mult = 0.9),
#          .by = c(respondent_id, employee_fte)) |> 
#   
#   ggplot(aes(y = y_pos, x = employee_fte)) +
#   
#   geom_point(colour = ca_cyan(), size = 2, alpha = 0.9) +
#   
#   scale_y_discrete(drop = FALSE) +
#   scale_x_continuous(breaks = seq(0, 1, 0.25), limits = 0:1) +
#   
#   theme(
#     axis.text.y = element_blank(),
#     axis.ticks.y = element_blank(),
#     panel.background = element_rect(fill = colorspace::lighten(ca_cyan(), 0.9),
#                                     colour = muted(ca_cyan(), 20, 50),
#                                     ),
#     strip.text = element_blank(),
#     panel.spacing.y = unit(.25, "lines"),
#     plot.caption = element_text(colour = "gray30"),
#     panel.grid.major.x = element_line(colour = "gray50",
#                                       linetype = "dotted"),
#     panel.grid.minor.x = element_blank()
#   ) +
#   
#   labs(y = "Churh plants",
#        x = "FTE (individual employees)",
#        caption = "Some FTE represent appx. % of split roles") + 
#   
#   facet_wrap(~respondent_id, dir = "v", ncol = 1, scales = "free_y",
#             strip.position = "left", drop = FALSE) +
#   ggtitle("Variation in Employees/FTE",
#           sub = "Across all church plants")

n_fte_resp <- sum(!is.na(scot$fte))

ggplot(scot, aes(x = fte)) + 
  geom_histogram(boundary = 0.01, binwidth = 0.5) +
  
  labs(
    title = "Total full-time-equivalent staff",
    x = "Total FTE",
    y = "Church plants") +
  
  scale_y_continuous(
    sec.axis = 
      sec_axis(
        transform = \(x) x / n_fte_resp,
        labels = label_percent(1),
        breaks = seq(0, 30, 10)/n_fte_resp))

scot |>
  mutate(fte_level = ceiling(fte * 2) / 2,
         label = str_c(fte_level - 0.5, " - ", fte_level),
         label = replace(label, label == "-0.5 - 0", "0"),
         label = ordered(label) |>
           fct_reorder(fte_level)) |>
  factor_table(label, x = _) |> 
  set_names("Total FTE", "Church plants", "%") |> 
  arrange(`Total FTE`) |> 
  kable()


```

## Mission focus

Which group of people is your church plant primarily aimed at/intended for? Please tick all that apply.

```{r plot_mission_focus}
mission_foci <- 
  scot |> 
  mutate(mission_focus_other = !is.na(mission_focus_other)) |> 
  filter(if_any(starts_with("mission_focus_"))) |>
  summarise(across(starts_with("mission_focus"),
                   list(prop = \(x) sum(x) / length(x),
                        count = \(x) sum(x)))) |> 
  pivot_longer(everything(),
               names_to = c("mission_focus", "stat"),
               names_pattern = "mission_focus_(.+)_(prop|count)") |>
  pivot_wider(names_from = stat) |> 
  mutate(mission_focus =
           ordered(mission_focus) |> 
           fct_reorder(-prop) |> 
           fct_relevel("none", after = Inf)) |> 
  arrange(-prop)

m_f_resp <- 
  scot |> 
  mutate(mission_focus_other = !is.na(mission_focus_other)) |> 
  filter(if_any(starts_with("mission_focus_"))) |> 
  nrow()

ggplot(mission_foci, aes(x = mission_focus, y = count)) + 
  geom_col() +
  
  scale_y_continuous(name = "Church plants",
                     breaks = seq(0,100,10),
                     sec.axis = sec_axis(
                       transform = \(x) x / m_f_resp,
                       labels = label_percent(),
                       breaks = seq(0,100,10)/m_f_resp,
                       name = "% responding plants"
                     )) +
  
  scale_x_discrete(name = "Mission focus",
                   labels = \(x) 
                   relabeller("dechurched" ~ "Church\naccess")(x) |> 
                     str_to_title()) +
  ggtitle("Most plants had a geographical mission focus",
          sub = "")

scot |> 
  select(respondent_id, starts_with("mission_focus")) |> 
  select(-mission_focus_none) |> 
  mutate(mission_focus_other = !is.na(mission_focus_other)) |> 
  filter(if_any(starts_with("mission_focus_"))) |> 
  pivot_longer(-respondent_id, names_to = "focus", values_to = "is") |> 
  summarise(n_foci = sum(is), .by = respondent_id) |>
  summarise(n_plants = n(), .by = n_foci) |>
  complete(n_foci = full_seq(n_foci, 1), fill = list(n_plants = 0)) |> 
  add_percent(n_plants) |> 
  rename(`Missional "foci"` = `n_foci`,
         `Church plants` = `n_plants`) |> 
  kable()

filter(scot, !is.na(mission_focus_other)) |> 
  pull(mission_focus_other)
```

Most church plants only identified one missional focus.

## % Non/dechurched

To the best of your knowledge, what percentage of your attenders come from \[non-churched\|de-churched\] backgrounds?

```{r pc_non_de}
scot <- mutate(scot, pc_non_de = pc_nonchurched + pc_dechurched)

scot$pc_nonchurched[scot$pc_non_de > 100] <- NA
scot$pc_dechurched[scot$pc_non_de > 100]  <- NA
scot$pc_non_de[scot$pc_non_de > 100] <- NA

non_de_resp <- sum(!is.na(scot$pc_non_de))

ggplot(filter(scot, pc_non_de <= 100), aes(x = pc_non_de)) +
  geom_histogram(binwidth = 10, boundary = 0.25) +
  
  scale_x_continuous(name = "Non/de-churched attenders",
                     breaks = seq(0, 100, 10), limits = c(0, NA),
                     labels = \(x) str_c(x, "%")) +
  scale_y_continuous(name = "Church plants", 
                     sec.axis = sec_axis(transform = \(y) y/non_de_resp,
                                         labels = label_percent(1),
                                         breaks = seq(0, 15, 5)/non_de_resp)) +
  ggtitle("% Non/de-churched attenders at each church plant") +
  
  theme(panel.grid.minor = element_blank()) 

non_de <- 
  pivot_longer(scot, starts_with("pc_") & !pc_non_de,
             names_to = "stat", values_to = "n",
             names_prefix = "pc_") |> 
  mutate(stat = str_to_title(stat))

n_non_de_resp <- ceiling(nrow(non_de)/2)
  
  ggplot(non_de, aes(x = n)) +
  
  geom_histogram(binwidth = 10, boundary = 0.5) +
  
  scale_x_continuous(breaks = seq(0, 100, 20),
                     name = "% Attenders",
                     labels = \(x) str_c(x, "%")) +
  
  scale_y_continuous(name = "Church plants",
                     breaks = seq(0, 100, 10),
                     sec.axis = 
                       sec_axis(
                         transform = \(x) x/n_non_de_resp,
                         labels = label_percent(1),
                         breaks = seq(0, 100, 10) / n_non_de_resp
                       )) +
  
  theme(panel.grid.minor.x = element_blank()) +
    
    ggtitle("% Non/Dechurched attenders across church plants") + 
  
  facet_wrap(~ stat, dir = "v")
```

## Faith commitments

How many faith commitments would you estimate your church has seen since it first started?

An overview of distribution for internal benefit

```{r faith_commitments}
library(ggrain)

n_resp_fc <- sum(!is.na(scot$total_faith_commitments))

scot |> 
  filter(!is.na(total_faith_commitments)) |> 
  mutate(
    faith_commitments =
    case_when(
      total_faith_commitments == 0 ~ "0",
      total_faith_commitments <= 10 ~ "1-10",
      total_faith_commitments <= 100 ~ "11-100",
      total_faith_commitments > 100  ~ "101+"
    ) |> 
      ordered() |> 
      fct_reorder(total_faith_commitments)) |> 
    group_by(faith_commitments) |> 
    count() |> 
  arrange(faith_commitments) |> 
  ungroup() |> 
  mutate(`%` = label_percent(1)(prop.table(n))) |> 
  rename("Faith commitments" = 1) |> 
  kable()

median(scot$total_faith_commitments, na.rm = TRUE)
```

## Success in reaching non/dechurched

Compared to your original aspirations, how successful/unsuccessful has the church plant been at reaching those who are non-churched or de-churched?

```{r success_reaching_non_de}

n_srn_resp <- sum(!is.na(scot$success_reaching_nondechurched))

filter(scot, !is.na(success_reaching_nondechurched)) |> 
ggplot(aes(x = success_reaching_nondechurched)) + 
  geom_bar() +
  scale_x_discrete(drop = FALSE, labels = \(x) str_replace(x, " ", "\n")) +
  scale_y_continuous(
    sec.axis = sec_axis(
      transform = \(x) x/n_srn_resp,
      labels = label_percent(1),
      breaks = seq(0, 30, 10)/n_srn_resp
    )
  ) +
  labs(
    y = "Church plants",
    x = "Success in reaching non/de-churched people",
    title = "Perceived missional impact",
    subtitle = "(Success compared to anticipated success)"
  )

scot |> 
  filter(!is.na(success_reaching_nondechurched)) |> 
  count_with_percent(success_reaching_nondechurched) |> 
  rename(`Success in reaching non/de-churched` = 1,
         `Church plants` = n) |> 
  complete(`Success in reaching non/de-churched`,
          fill = list(`Church plants` = 0, `%` = "0%")) |> 
  kable()
```

### Difference services

On a scale from 1 to 100, how different are your church's regular services from those of your sending church? (0: absolutely identical, 100: completely re-imagined)

```{r difference_services}
# scot$how_different_from_sending_services

median_send_diff <- median(scot$how_different_from_sending_services, 
                           na.rm = TRUE)
ggplot(scot, aes(x = how_different_from_sending_services)) + 
  geom_density(fill = ca_cyan(), col = ca_cyan(), 
               alpha = 0.5) +
  
  geom_vline(xintercept = median_send_diff, linetype = "dashed") + 
  
  scale_x_continuous(lim = c(0,100)) +

  
  labs(
    y = "Density",
    x = "Difference from sending church service",
    title = "How different are services to sending church?",
    subtitle = "Density visualisation of scores from 0 - 100"
  ) +
  
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank()
  )

scot |> 
  rename(how = how_different_from_sending_services) |> 
  filter(!is.na(how)) |> 
  mutate(
    how = 
      case_when(
        how == 0 ~ "0%",
        how <= 25 ~ "1-25%",
        how <= 50 ~ "26-50%",
        how <= 75 ~ "51-75%",
        how <= 99 ~ "76-99%",
        how == 100 ~ "100%"
      ) |> 
      ordered() |> 
      fct_reorder(how)
  ) |> 
  group_by(how) |> 
  count() |> 
  ungroup() |> 
  transmute(
    `% Difference` = how,
    n = n,
    `%` = label_percent(1)(prop.table(n))) |> 
  arrange(`% Difference`) |> 
  kable()
```

## Extent to which church plants have had to develop new missional practises

To what extent has the church plant had to develop new practices in order to reach and disciple people who do not already go to church?

```{r develop_practises}
n_prac_resp <- sum(!is.na(scot$extent_develop_new_missional_practices))

filter(scot, !is.na(extent_develop_new_missional_practices)) |> 
ggplot(aes(x = extent_develop_new_missional_practices)) +
  geom_bar() +
  
  scale_x_discrete(drop = FALSE) +
  scale_y_continuous(
    sec.axis = sec_axis(
      transform = \(x) x/n_prac_resp,
      labels = label_percent(),
      breaks = seq(0, 30, 10)/n_prac_resp
    )
  ) +
  labs(
    title = "Extent to which new missional practises have been developed",
    y = "Church plants",
    x = "Extent"
  )

scot |> 
  filter(!is.na(extent_develop_new_missional_practices)) |> 
  count_with_percent(extent_develop_new_missional_practices) |> 
  complete(extent_develop_new_missional_practices,
           fill = list(n = 0, `%` = "0%")) |> 
  rename(
    "Extent of development of new missional practices" = 1,
    "Church plants" = 2
  ) |> 
  kable()
```

## Various freetext fields

Based on your experiences, what would you say are the key factors that help determine whether a church plant thrives and flourishes?

### Flourish factors

```{r flourish_factors}
#| output: asis

bullets <- function(x){
  str_c("* ", x, collapse = "\n")
}

scot |> 
  pivot_longer(starts_with("flourish_")) |> 
  filter(!is.na(value)) |> 
  pull(value) |> 
  bullets() |> 
  cat()
```

### Inhibit factors

Based on your experiences, what would you say are the key factors that inhibit a church plant's growth and development?

```{r inhibit_factors}
#| output: asis 
scot |> 
  pivot_longer(starts_with("inhibit_")) |> 
  filter(!is.na(value)) |> 
  pull(value) |> 
  bullets() |> 
  cat()
```

### How engage non/de churched

```{r how_enage_non_de}
scot |> 
  filter(!is.na(how_engage_nondechurched)) |> 
  select(how_engage_nondechurched) |> 
  kable()
```

## Measures of central tendency

```{r measures_central_tendency}
scot <- 
  mutate(scot, total_current_attenders =
           current_attenders_children + current_attenders_adult) |> 
  relocate(total_current_attenders, .after = current_attenders_children)

scot <- relocate(scot, employees, fte, .after = staff_member_8_fte)
scot <- relocate(scot, launch_date, .after = start_month)
scot <- relocate(scot, pc_non_de, .after = pc_dechurched)

scot <- select(scot, where(\(x) !all(is.na(x))))

scot |> 
  summarise(across(where(is.numeric),
            list(
              sum = \(x) sum(x, na.rm = TRUE),
              mean =   \(x) mean(x, na.rm = TRUE),
              median = \(x) median(x, na.rm = TRUE),
              max =    \(x) max(x, na.rm = TRUE),
              min =    \(x) min(x, na.rm = TRUE)),
            .names = "{.col}//{.fn}")) |> 
  pivot_longer(everything(), names_sep = "//",
               names_to = c("variable", "statistic")) |> 
  pivot_wider(names_from = "statistic") |> 
  mutate(across(where(is.numeric), \(x) round(x, 2))) |> 
  kable()
```

## Overall non-dechurched attenders

```{r overall_non_de}
scot |> 
  mutate(non = 0.01 * pc_nonchurched * total_current_attenders,
         de  = 0.01 * pc_dechurched  * total_current_attenders) |> 
  summarise(across(c(non, de, total_current_attenders),
                   \(x) sum(x, na.rm = TRUE))) |> 
  mutate(pcn = label_percent()(non / total_current_attenders),
         pcd = label_percent()(de / total_current_attenders))

scot |> 
  group_by(pc_nonchurched == 0 & pc_dechurched == 0) |> 
  count()

2/84

non_de_labels <-
  scot |> 
  mutate(across(c(pc_nonchurched, pc_dechurched),
                \(x){
                  lab <- ceiling(x / 10) * 10
                  
                  lab <- str_c(lab - 10, " - ", lab)
                  
                  lab <-
                    replace(lab, lab == "-10 - 0", "0") |> 
                    str_c("%")
                    
                    ordered(lab) |>
                    fct_reorder(x)
                }))

non_de_labels |> 
  group_by(pc_nonchurched) |> 
  count() |> 
  ungroup() |> 
  mutate(prop = n/sum(n)) |> 
  filter(!is.na(pc_nonchurched)) |> 
  transmute(`% non-churched` = pc_nonchurched,
            `Church plants` = n, 
            `%` = label_percent(1)(prop)) |> 
  kable()

```

```{r de_table}
  

non_de_labels |> 
  group_by(pc_dechurched) |> 
  count() |> 
  ungroup() |> 
  mutate(prop = n/sum(n)) |> 
  filter(!is.na(pc_dechurched)) |> 
  transmute(`% de-churched` = pc_dechurched,
            `Church plants` = n, 
            `%` = label_percent(1)(prop)) |> 
  kable()
```

## Converts per year

```{r converts_per_year}
scot <- 
  mutate(scot,
         years_running = as.numeric((date(start_date) - launch_date)/365.25),
         years_running = replace(years_running, years_running < 0, 0),
         commitments_per_year = total_faith_commitments/years_running)

scoot <- filter(scot, total_faith_commitments < 999)
 
ggplot(scoot, aes(x = total_faith_commitments)) + 
  geom_density()

ggplot(scoot, aes(x = commitments_per_year)) + 
  geom_density()

median(scot$commitments_per_year, na.rm = TRUE)

levels_after <- function(f, level){
  lf <- levels(f)
  n <- which(lf == level)
  lf[n:length(lf)]
}

filter(scot, !is.na(commitments_per_year), !is.infinite(commitments_per_year)) |> 
  mutate(cpy = range_labels(commitments_per_year, width = 1, continuous = 0.1),
         cpy = fct_collapse(cpy, "4+" = levels_after(cpy, "4.1 - 5")) |>
           fct_explicit_na("4+")) |> 
  group_by(cpy) |> 
  count() |> 
  ungroup() |> 
  mutate(`%` = label_percent()(prop.table(n))) |> 
  set_names("Faith commitments per year", "Church plants", "%") |> 
  kable()
```


## Save

```{r save}
saveRDS(scot, here("data", "output-02_report-data.RDS"))
```
